usecaseのディレクトリ構成を変更
