+++ src/app/Usecase/SignIn/SignInInput.php
@@ -1,26 +1,15 @@
 <?php
-namespace App\UseCase\UseCaseInput;
+namespace App\UseCase\SignIn;
 
-final class SignupInput
+final class SignInInput
 {
-    private $name;
     private $email;
     private $password;
 
-    public function __construct(string $name, string $email, string $password)
+    public function __construct(string $email, string $password)
     {
-        $this->name = $name;
         $this->email = $email;
         $this->password = $password;
-    }
-
-    /**
-     * name入力を取得
-     * @return string
-     */
-    public function getName(): string
-    {
-        return $this->name;
     }
 
     /**
